refactor: migrate the AuthController to the user module
